Pivot to index segments by both center and target
The great wonder of writing documentation once again asserts itself!
Only while writing up the documentation — frighteningly enough, the last
step before release — did I realize that the API was dangerously
misleading because it let you just ask the position of a target, without
any context as to where the measurement was from.

So the quick go-to indexing that the SPK object supports — which I now
like so very much that I have promoted it to the __getitem__ of the
kernel object — now makes the user name both the center and the target
that they want a vector between.