Switch from one big mmap() to an mmap() per array
Bringing the whole file in at once was offering the error

OSError: [Errno 12] Cannot allocate memory

in experiments I was doing with Skyfield with Jovian moons, because of
the large size of jup310.bsp plus some other data that I was using.  So
with this change, jplephem uses mmap() separately for each segment the
user needs, making it possible to map only one — or a few — segments of
a file at a time, discarding older ones as newer ones are needed and
limiting the amount of simultaneously mapped memory required.